completed fuzzy search, decimal conversion, and part of companies(transactions don't reduce credit amt)

+++ elixir/lib/homework_web/schemas/transactions_schema.ex
@@ -9,7 +9,7 @@ defmodule HomeworkWeb.Schemas.TransactionsSchema do
   object :transaction do
     field(:id, non_null(:id))
     field(:user_id, :id)
-    field(:amount, :integer)
+    field(:amount, :float)
     field(:credit, :boolean)
     field(:debit, :boolean)
     field(:description, :string)
@@ -30,9 +30,10 @@ defmodule HomeworkWeb.Schemas.TransactionsSchema do
     @desc "Create a new transaction"
     field :create_transaction, :transaction do
       arg(:user_id, non_null(:id))
+      # arg(:company_id, non_null(:id))
       arg(:merchant_id, non_null(:id))
       @desc "amount is in cents"
-      arg(:amount, non_null(:integer))
+      arg(:amount, non_null(:float))
       arg(:credit, non_null(:boolean))
       arg(:debit, non_null(:boolean))
       arg(:description, non_null(:string))
